% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMonths.R
\name{AddMonths}
\alias{AddMonths}
\title{Add months to a date.}
\usage{
AddMonths(date, n)
}
\arguments{
\item{date}{An R-date.}

\item{n}{An integer.}
}
\value{
The date shifted by n months.
}
\description{
\code{AddMonths} increases the month part of a date by n months (as opposed to adding a certain number of days). This works
 for both positive and negative n. It requires the lubridate package.
}
\examples{
 AddMonths(as.Date(c("2017-01-01")), 12)
 AddMonths(as.Date(c("2017-01-01")), -12)
}

