% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregatePrepayData.R
\name{AggregatePrepaymentData}
\alias{AggregatePrepaymentData}
\title{Aggregate prepayment data along a specified dimension}
\usage{
AggregatePrepaymentData(data, act, pred, col.name, configData)
}
\arguments{
\item{data}{A data frame with prepayment data}

\item{act}{actual prepayments (in smm)}

\item{pred}{model forecasted prepayments (in smm)}

\item{col.name}{column to aggregate over}

\item{configData}{configuration data}
}
\value{
A list containing actual and projected prepays (in cpr) when aggregated along the specified dimension. 
Additionally, the act/proj ratio, balances and column values associated with the gridding of the dimension are also
provided.
}
\description{
\code{AggregatePrepaymentData} calculates the weighted average prepayment rate (actual and model) associated with a 
 specific bucketing of the specified dimension. It requires the data.table package.
}
\examples{
\dontrun{
AggregatePrepaymentData(data=ppdata, act=act.smm, pred=mdl.smm, col.name="asofdate")
AggregatePrepaymentData(data=ppdata, act=act.smm, pred=mdl.smm, col.name="burnout")
}

}

