% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildSpline.R
\name{BuildSpline}
\alias{BuildSpline}
\title{Construct a cubic, Hermite, or smooth spline function from given data points}
\usage{
BuildSpline(a, b, x, weights = NULL, method = NULL)
}
\arguments{
\item{a}{x-values for knot points}

\item{b}{y-values for knot points}

\item{x}{interpolation values for spline}

\item{weights}{Use weights in spline construction}

\item{method}{What spline to build. The three choices right now are "natural", "monoH.FC" and "smooth".}
}
\value{
data frame with spline values
}
\description{
\code{BuildSpline} Given data points in (a, b) format, construct a cubic, Hermite or smooth spline function. 
This function is then used to interpolate for a more comprehensive set of x-coordinates.
}
\examples{
 n <- 20
 set.seed(11)
 a <- sort(runif(n)); b <- cumsum(abs(rnorm(n))); x <- round(rnorm(30), 1)
 BuildSpline(a, b, x, method="monoH.FC")
 
}

