% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateEQLODMultiplier.R
\name{CalculateEQLODMultiplier}
\alias{CalculateEQLODMultiplier}
\title{Modify eqlod multiplier as a function of the current LTV.}
\usage{
CalculateEQLODMultiplier(origMult, cltv)
}
\arguments{
\item{origMult}{original eqlod multiplier.}

\item{cltv}{a vector of cltv values.}
}
\value{
eqlod multiplier.
}
\description{
\code{CalculateEQLODMultiplier} sets the previously calculated eqlod multiplier to 1.0 whenever the cltv is greater
 than 100%.
}
\examples{
 CalculateEQLODMultiplier(origMult, cltv)
 
}

