% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateNewBurnout.R
\name{CalculateNewBurnout}
\alias{CalculateNewBurnout}
\title{Calculate burnout as a function of cumulative incentive and current loan size.}
\usage{
CalculateNewBurnout(F, L)
}
\arguments{
\item{F}{a vector of values for F(I).}

\item{L}{a vector of loan sizes corresponding to I.}
}
\value{
Burnout multiplier.
}
\description{
\code{CalculateNewBurnout} estimates burnout as a function of cumulative incentive and loan size. It assumes that
 burnout is given by B(I,L) = 1-[(1 - F(I)) * G(L)] where:
I = cumulative incentive
L = current loan size
F is a decreasing function with F(0) = 1. We take F to be the usual burnout multiplier as a function of cumulative
incentive. G is an increasing function of WACLS.
Note that B(0, L) = 1 and that B(I,L) = F(I) if G(L) = 1 for all L.
}
\examples{
\dontrun{
 CalculateNewBurnout(F, L)
 }
 
}

