% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRefiCosts.R
\name{CalculateRefiCosts}
\alias{CalculateRefiCosts}
\title{Calculate refi costs for different loan/property attributes.}
\usage{
CalculateRefiCosts(ppdata, configData)
}
\arguments{
\item{ppdata}{Data frame of prepayment observations}
}
\value{
ppdata Prepayment data frame with an extra column for refi-related costs
}
\description{
\code{CalculateRefiCosts} estimates the cost of a refinancing based on different loan and/or property attributes. 
 In particular, these costs can vary by loan size, whether the property is a second home, and/or if it is an investor
 property, among other attributes.
}
\details{
TO-DO
}
\examples{
 CalculateRefiCosts(ppdata=ppdata)
  
}

