% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Instantiate variables that parameterize batonutil workflows}
\usage{
config(writeJSON = "no")
}
\description{
\code{config} creates a JSON file that stores key variables that are used in batonutil.
 
 @param writeJSON Option to write configuration JSON to a file
 
 @return JSON object with configuration parameters
 
 @examples
 \dontrun{
 config()
 config(writeJSON="yes")
 }
}

