% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{imprtr}
\alias{imprtr}
\title{Read in prepayment data from csv files}
\usage{
imprtr(infile = NA)
}
\arguments{
\item{infile}{Input filename and location}
}
\value{
ppdata Prepayment data as a data frame
}
\description{
\code{imprtr} reads in various prepayment data sets.  It also manages naming conventions and type conversions for various columns.
See also \code{\link{tfrmr}} for functions that perform additional data munging.
}
\details{
\code{imprtr} reads in various aggregations (replines) of Agency MBS pool- and loan-level prepayment data.
The input file name and location is passed in as an argument.
}
\examples{
 imprtr(infile=RschGridFile)
 
}

