% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckMontone.R
\name{isMonotone}
\alias{isMonotone}
\title{Check to see if a vector is monotone increasing or decreasing}
\usage{
isMonotone(x)
}
\arguments{
\item{x}{numeric or a complex object that can be coerced into a numeric}
}
\value{
Boolean indicating wheter the vector is monotone or not.
}
\description{
\code{isMonotone} Check to see if a vector is monotone increasing or decreasing; useful in ensuring that we are selecting the
correct spline building utilities some of which will only work with monotone inputs.
}
\examples{
 set.seed(11)
 isMonotone(rnorm(10))
 
}

