% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelErrorStats.R
\name{ModelErrorStats}
\alias{ModelErrorStats}
\title{RMSE with option to use weights for different points and to also allow computation of an adjusted RMSE}
\usage{
ModelErrorStats(actual, pred, weights = NULL, totalparams = 0,
  freeparams = 0)
}
\arguments{
\item{actual}{observed prepayment rates}

\item{pred}{model predictions}

\item{weights}{weights for the various observations}

\item{totalparams}{not used}

\item{freeparams}{not used}
}
\value{
List with error statistics
}
\description{
\code{ModelErrorStats} calculates the RMSE given a vector of actual prepayments and a vector of projected prepayments.
}
\examples{
\dontrun{
 ModelErrorStats(actual=actual, pred=pred, weights)
 }
 
}

