% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelProjections.R
\name{ModelProj}
\alias{ModelProj}
\title{Generate prepayment forecasts}
\usage{
ModelProj(configData, modelVersions = NA, ppdata)
}
\arguments{
\item{configData}{Configuration data for prepayment model}

\item{modelVersions}{Release numbers for different versions of the prepayment model}

\item{ppdata}{data frame containing indicative data for the loans/pools/replines that we want projections for}
}
\description{
\code{ModelProj} Generate model and submodel forecasts based on indicative data from a data frame.
The forecasts are added as columns to the data frame.
}
\examples{
\dontrun{
 ModelProj(configData=configData, modelVersions=c("v2.42", "v3.00"), ppdata=ppdata)
 }
 
}

