% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseModelJSON.R
\name{ParseModelJSON}
\alias{ParseModelJSON}
\title{Parse model parameters from JSON file}
\usage{
ParseModelJSON(configData, modelVersion = NA)
}
\arguments{
\item{modelVersion}{Release number for the prepayment model}

\item{coll}{collateral type}

\item{gnmaProgram}{GNMA program (ex., FHA) to read submodel JSON files for}
}
\value{
List whose elements consists of individual submodel JSONs for collateral type and model version
}
\description{
\code{ParseModelJSON} reads in JSON files that store splines corresponding to the various submodels for a given
collateral type and for a specified version of the model.
}
\examples{
\dontrun{
ParseModelJSON(configData=configData, modelVersion="v2.00")
}
 
}

