% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictDflt.R
\name{PredictDflt}
\alias{PredictDflt}
\title{Generate forecasts for the default submodel model. Other submodels have a separate function.}
\usage{
PredictDflt(curr.spl = NA, delq.spl = NA, data, modelVersion = "")
}
\arguments{
\item{curr.spl}{Spline coefficients for current loans}

\item{delq.spl}{Spline coefficients for delinquent loans}

\item{data}{data frame containing indicative data for the loans/pools/replines that we want projections for}

\item{version}{Model version. Allows flexibility in changing model specifications for different versions.}
}
\value{
List with submodel predictions
}
\description{
\code{PredictDflt} Generate default forecasts based on indicative data from a data frame.
The forecasts are added as columns to the data frame. Other submodels are driven by PredictSubmodel.
}
\examples{
\dontrun{
 PredictDflt(curr.spl=curr.spl, delq.spl=NA, data=ppdata, modelVersion="v4.00")
 }
 
}

