% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictSubmodel.R
\name{PredictSubmodel}
\alias{PredictSubmodel}
\title{Generate forecasts for a submodel of the prepayment model. Default submodel has a separate function.}
\usage{
PredictSubmodel(coll, submodel = NA, submdl.spl, data, seasadj = FALSE,
  version = "")
}
\arguments{
\item{coll}{Collateral type for prepayment model}

\item{submodel}{Submodel for prepayment model: turn, refi, cout, dflt}

\item{submdl.spl}{Spline coefficients for the various variables in the model}

\item{seasadj}{Produce seasonally-adjusted forecasts if this is TRUE}

\item{version}{Model version. Allows flexibility in changing model specifications for different versions.}

\item{ppdata}{data frame containing indicative data for the loans/pools/replines that we want projections for}
}
\value{
List with submodel predictions
}
\description{
\code{PredictSubmodel} Generate submodel forecasts based on indicative data from a data frame.
The forecasts are added as columns to the data frame. Default submodel is driven by PredictDflt.
}
\examples{
\dontrun{
 PredictSubmodel(submodel="turn", submdl.spl=turn.mdl, data=ppdata, version="v4.00")
 }
 
}

