% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepayComponents.R
\name{PrepayComp}
\alias{PrepayComp}
\title{Decompose actual prepayment rate into subcomponents by prepayment type}
\usage{
PrepayComp(submodel, modelVersion, ppdata)
}
\arguments{
\item{submodel}{submodel type}

\item{modelVersion}{Prepayment model version}

\item{ppdata}{data frame containing indicative data for the loans/pools/replines that we want projections for}
}
\value{
Prepayment model dataframe with added columns
}
\description{
\code{PrepayComp} Decomposes the actual prepayment rate into its constituents (curtailments, default, ...) by
using the prepayment model. The actual components are added as columns to the data frame. Note that the decomposition
depends on the choice of prepayment model
}
\examples{
\dontrun{
 PrepayComp(submodel="turn", ppdata=ppdata, modelVersion)
 }
 
}

