% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasonalAdjustment.R
\name{SeasonalAdjustment}
\alias{SeasonalAdjustment}
\title{Generate seasonal factors for different months}
\usage{
SeasonalAdjustment(seasonals, months)
}
\arguments{
\item{seasonals}{seasonal multipliers for different calendar months in the year}

\item{months}{vector of months}
}
\value{
Vector of seasonal multipliers corresponds to the different months
}
\description{
\code{SeasonalAdjustment} reads in seasonal multipliers corresponding to different months and then calculates
the appropriate multiplier for a vector of months
}
\examples{
\dontrun{
SeasonalAdjustment(seasonals=seasonals, months=months)
}
 
}

