% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{tfrmr}
\alias{tfrmr}
\title{Transform prepayment data}
\usage{
tfrmr(configData, ppdata)
}
\arguments{
\item{configData}{configuration data for the prepayment model}

\item{ppdata}{Data frame of prepayment data}
}
\value{
ppdata Transformed prepayment data frame
}
\description{
\code{tfrmr} adds new variables that are combinations of existing columns. It also filters the prepayment data frame
if necessary.
}
\details{
(1) Modify the incentive by a cost function. 
(2) Add weights that correspond to the relative balance of each data point. 
(3) Apply data filters where appropriate
}
\examples{
 tfrmr(configData=configData, ppdata=ppdata)
 
}

