% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr.R
\name{trackr}
\alias{trackr}
\title{Track actual versus projected prepayments}
\usage{
trackr(configData, ticker = "", ppdata)
}
\arguments{
\item{configData}{List of configuration data for the fitting process}

\item{ticker}{ticker for collateral  Examples: (\code{"FGLMC"}, \code{"FGT6"}, \code{"FGU6"}, \code{"FGU9"})}

\item{ppdata}{Prepayment data frame}
}
\value{
Prepayment data frame
}
\description{
\code{trackr} tracks actual versus projected prepayments based on user-specified cohorts
}
\examples{
\dontrun{
 trackr(configData=configData, ppdata=ppdata)
 }
 
}

